% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_df.R
\name{add_to_df}
\alias{add_to_df}
\title{add_to_df}
\usage{
add_to_df(df, v)
}
\arguments{
\item{df}{a data frame where we want to include a columns probably with dimensions different than the ones the data.frame allows}

\item{v}{a vector containing values to be added as column}
}
\description{
Adds a column with number of elements different than the number of rows of the data frame we want to fill the column. cbind does not work in this scenario.
}
