% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assure_cor.R
\name{assure_corr}
\alias{assure_corr}
\title{assure_corr}
\usage{
assure_corr(n_var = n_var, corr = corr)
}
\arguments{
\item{n_var}{number of variables inside a module}

\item{corr}{a matrix with values between -1 and 1 to be transforme in a proper correlation matrix}
}
\description{
assure_corr
}
