% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_trig_dis.R
\name{calculate_trig_dis}
\alias{calculate_trig_dis}
\title{calculate_trig_dis}
\usage{
calculate_trig_dis(known_cor)
}
\arguments{
\item{known_cor}{Values of known correlation to be transformed into trigonometric distances}
}
\value{
calculated trigometric distances
}
\description{
Calculates trigonometric distances based on know correlations
}
\examples{
 calculate_trig_dis(c(0.80,-0.99))
}
