% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_templates.R
\name{create_dap}
\alias{create_dap}
\title{Create a Data Analysis Project}
\usage{
create_dap(
  dir,
  project_type = "report",
  git_activate = FALSE,
  renv_activate = FALSE
)
}
\arguments{
\item{dir}{Directory for data analysis project.}

\item{project_type}{The type of project is indicated.
The options are "reporting" or "production".}

\item{git_activate}{Configure the project so that it
can use git.}

\item{renv_activate}{Configure the project so that it
can use renv.}
}
\description{
Create a skeleton for a data analysis project in 2
formats: report, Report-centric (Rmd's); production,
Modularised (separate scripts and Rmd's). Use the
`create_dap()` function for this.
}
\note{
The `dir` parameter is required for the creation
  of the project. By default a project of type report is configured.
}
\examples{
\dontrun{
library(innovar)
create_dap("Epi_project", "report")
}
}
