% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_levels.R
\name{cut_levels}
\alias{cut_levels}
\title{cut_levels}
\usage{
cut_levels(
  categorical = soil_cat,
  cat_inc_all_levels = c("siteid", "yes_no_sample"),
  levels_inc_all = list(siteid = c(1:16), yes_no_sample = c(1:2)),
  continuous_to_cut = sim_data_cop$simulated,
  corrs = soil_corr
)
}
\arguments{
\item{categorical}{categorical variables in a module}

\item{cat_inc_all_levels}{categorical variables which all levels must be in final simulations}

\item{levels_inc_all}{levels of above}

\item{continuous_to_cut}{a matrix containing values simulated from copulas}

\item{corrs}{a coorelation matrix}
}
\description{
This function creates a vector of levels based on a continuous vector. It allows to force all the levels desired for some of the variables respecting the correlation of the variable with other variable in a module.
}
