% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ee_std}
\alias{ee_std}
\title{Set of function to zonal statistic}
\usage{
ee_std(x, y, by = 1000)
}
\arguments{
\item{x}{image of type Image o Image Collection.}

\item{y}{region of type feacture or feature collection.}

\item{by}{a limit of pass by polygon.}
}
\value{
a object sf with the standard deviation values.
}
\description{
Set of function to zonal statistic
}
