% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ee_sum}
\alias{ee_sum}
\title{Set of function to zonal statistic}
\usage{
ee_sum(x, y, by = 1000)
}
\arguments{
\item{x}{image of type Image o Image Collection.}

\item{y}{region of type feature o feature collection.}

\item{by}{a limit of pass by polygon.}
}
\value{
a object sf with the sum values.
}
\description{
Set of function to zonal statistic
}
