% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export CSV file to MongoDB}
\usage{
export(filepath, dbname, user, password, interpreter = NULL)
}
\arguments{
\item{filepath}{CSV file path.}

\item{dbname}{Database name.}

\item{user}{User name.}

\item{password}{Password.}

\item{interpreter}{Path of the Python interpreter (optional)}
}
\value{
Confirmation message if exporting was successful. Else an error.
}
\description{
This a Python wrapper of the \code{insert_many()} method of package pymongo to insert a list of documents as a collection in a database in MongoDB.
}
\examples{
\dontrun{
library(innovar)
data("cars")
write.csv(cars, "cars.csv", row.names = FALSE)
dbname = "<database-name>"
user <- Sys.getenv("USER")
password <- Sys.getenv("PASSWORD")
export("cars.csv", dbname, user, password)
}
}
