% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_soil_gen.R
\docType{data}
\name{final_soil_gen}
\alias{final_soil_gen}
\title{final_soil_gen}
\format{
\subsection{\code{final_soil_gen}}{

A data frame containing large number of soil attributes
\describe{
\item{trait_i}{genomic expression of trait i}
\item{soil attributes}{ soil attributes measured}

...
}
}
}
\source{
<the author>
}
\usage{
final_soil_gen
}
\description{
A dataset containing genomic and soil simulated information
}
\keyword{datasets}
