% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_admin.div.R
\name{gen_admin_div}
\alias{gen_admin_div}
\title{Create categories for adiministrative divisions (Peru)}
\usage{
gen_admin_div(x, method = "Traditional")
}
\arguments{
\item{x}{x is a list/vector with one of the 25 regions/departments names of
Peru (first-level administrative subdivisions).}

\item{method}{defines the type of natural, geographical, or administrative
  division that should be returned. Possible values are:

  - Traditional: Coastal, Andes Mountains, Amazon Jungle, or Lima & Callao
  - TC: (Lima no incluye Huarua/Cañete)\href{https://andina.pe/agencia/noticia-tc-crea-cinco-macro-regiones-para-facilitar-acceso-a-justicia-constitucional-604658.aspx}{Administratives Macro Regions of the Constitutional Tribunal}
  - Cardinal: North, South, Amazon Jungle, or Lima & Callao
  - Trad_Cardinal:  Coastal and Andes Mountains are divided into North and South (e.g. South Andes Mountains)}
}
\value{
Converts characters where it detects regions of Peru with the resulting divisions specified.
}
\description{
Function 'gen_admin_div' returns the natural, geographical, or
  administrative division of Peru that includes the region, province or district
  provided to the function.
}
