% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_att.R
\docType{data}
\name{gen_att}
\alias{gen_att}
\title{Genomic attributes considered}
\format{
\subsection{\code{gen_att}}{

A data frame 100 markers and its levels of expression (0,1,2) and trial sites levels:
\describe{
\item{site}{Site IDs}
\item{trait_i}{column containing the levels of expression of the trait i}

...
}
}
}
\source{
<the author>
}
\usage{
gen_att
}
\description{
A dataset containing attributes used to simulate genomic information
}
\keyword{datasets}
