% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gamma_delta.R
\name{generate_gamma_delta}
\alias{generate_gamma_delta}
\title{generate_gamma_delta}
\usage{
generate_gamma_delta(INDEX, Q)
}
\arguments{
\item{INDEX}{describes number of genotypes I and environments J}

\item{Q}{number of componentes captured by the bi linear term. (number of lambdas)}
}
\value{
a matrix n_genotypes by Q or n_environments by Q
}
\description{
Generates parameters lambda and delta of Bilinear terms. Imported from earlier versions of AMBARTI package.
}
\examples{
n_genotypes<-4
n_environments<-3
generate_gamma_delta(I=n_genotypes,Q=2)
generate_gamma_delta(I=n_environments,Q=2)
}
