% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gamma_delta_sigmas.R
\name{generate_gamma_delta_sigmas}
\alias{generate_gamma_delta_sigmas}
\title{generate_gamma_delta_sigmas}
\usage{
generate_gamma_delta_sigmas(INDEX, Q, sigma)
}
\arguments{
\item{INDEX}{describes number of genotypes I and environments J}

\item{Q}{number of componentes captured by the bi linear term. (number of lambdas)}

\item{sigma}{sigma associated with the parameters}
}
\value{
a matrix n_genotypes by Q or n_environments by Q
}
\description{
Generates parameters lambda and delta of Bilinear terms.
}
\examples{
n_genotypes<-4
n_environments<-3
generate_gamma_delta_sigmas(I=n_genotypes,Q=2,sigma=2)
generate_gamma_delta_sigmas(I=n_environments,Q=2,sigma=1)
}
