% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate.R
\name{get_climate}
\alias{get_climate}
\title{Extract climate data of TerraClimate}
\usage{
get_climate(year, bands, region, fun = "mean")
}
\arguments{
\item{year}{is date format for extract the variable}

\item{bands}{is a specif variable of climate}

\item{region}{is a sf object}

\item{fun}{is a type of fucntion for extract statistic zonal ('mean','median','sum','count')}
}
\value{
a sf object with the new variables
}
\description{
A function that extract a time series of climate variables
}
\details{
Name of some bands
\itemize{
\item pr: Precipitation accumulation in mm
\item ro: Runoff, derived using a one-dimensional soil water balance model in mm
\item tmmn: Minimum temperature in C°
\item tmmx: Maximum temperature in C°
}
}
