% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_co.R
\name{get_co}
\alias{get_co}
\title{Extract Carbon Monoxide data of Sentinel5}
\usage{
get_co(from, to, band, region, fun = "max", scale = 1000)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{band}{name of band.}

\item{region}{is a feature or feature collection.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance, first).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variables.
}
\description{
A function that extract a time series of carbon monoxide (2018-11-22T12:00:13Z - 2022-05-15T00:00:00).
}
\details{
Name of some bands.
\itemize{
\item \bold{CO_column_number_density (mol/m²):} Vertically integrated CO column density.
\item \bold{H2O_column_number_density (mol/m²):} Water vapor column.
\item \bold{cloud_height (m):} Scattering layer height.
\item \bold{sensor_altitude (m):} Altitude of the satellite with respect to the geodetic sub-satellite point (WGS84).
\item \bold{sensor_azimuth_angle (degrees):} Azimuth angle of the satellite at the ground pixel location (WGS84); angle measured East-of-North.
\item \bold{sensor_zenith_angle (degrees):} Zenith angle of the satellite at the ground pixel location (WGS84); angle measured away from the vertical.
\item \bold{solar_azimuth_angle (degrees)} Azimuth angle of the Sun at the ground pixel location (WGS84); angle measured East-of-North.
\item \bold{solar_zenith_angle (degrees):} Zenith angle of the satellite at the ground pixel location (WGS84); angle measured away from the vertical.
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
# 2. Extracting climate information
data <- region_ee \%>\% get_co(
  from = "2019-02-01", to = "2019-12-31",
  band = "CO_column_number_density", fun = "max")
}
}
