% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_def.R
\name{get_def}
\alias{get_def}
\title{Extract deforestation area data from Hansen}
\usage{
get_def(from, to, region, scale = 100)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{region}{region and object sf.}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variables.
}
\description{
A function that extract deforestation area data of the year \bold{2001-2019}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region, id = 'distr' ,simplify = 1000)

# 2. Extracting deforestation area data
data <- region_ee \%>\%
get_def(
 from = '2001-01-01',
 to = '2005-12-31',
 region = region_ee,
 scale = 30
)
}
}
