% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_etp.R
\name{get_etp}
\alias{get_etp}
\title{Extract Evapotranspiration data of MODIS}
\usage{
get_etp(year, region)
}
\arguments{
\item{year}{is date format for extract the variable}

\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract a ETP time series of MODIS
}
