% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fldas.R
\name{get_fldas}
\alias{get_fldas}
\title{Extract climate data of Famine Early Warning Systems Network (FEWS NET) Land Data Assimilation System ()}
\usage{
get_fldas(from, to, by, band, region, fun = "mean", scale = 1000)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{by}{two types of increment of the sequence by \bold{month} and \bold{year}.}

\item{band}{name of band.}

\item{region}{is a feature or feature collection.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance, first).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variables.
}
\description{
A function that extract a time series of climate variables (1982-01-01T00:00:00Z–2022-04-01T00:00:00).
}
\details{
Name of some bands.
\itemize{
\item \bold{Evap_tavg (kg m-2 s-1):} Evapotranspiration.
\item \bold{LWdown_f_tavg (W m-2):} Downward longwave radiation flux .
\item \bold{Lwnet_tavg (W m-2):} Net longwave radiation flux.
\item \bold{Psurf_f_tavg (Pa):} Surface pressure.
\item \bold{Qair_f_tavg (kg kg-1):} Specific humidity.
\item \bold{Qg_tavg (W m-2):} Soil heat flux.
\item \bold{Qh_tavg (W m-2):} Sensible heat net flux.
\item \bold{Qle_tavg (W m-2):} Latent heat net flux.
\item \bold{Qs_tavg (kg m-2 s-1):} Storm surface runoff.
\item \bold{Qsb_tavg (kg m-2 s-1):} Baseflow-groundwater runoff.
\item \bold{RadT_tavg (K):} Surface radiative temperature.
\item \bold{Rainf_f_tavg (kg m-2 s-1):} Total precipitation rate.
\item \bold{SnowCover_inst :} Snow cover fraction.
\item \bold{SnowDepth_inst (m):} Snow depth.
\item \bold{Snowf_tavg (kg m-2 s-1):} Snowfall rate.
\item \bold{SoilMoi00_10cm_tavg (m^3 m-3):}Soil moisture (0 - 10 cm underground).
\item \bold{SoilMoi10_40cm_tavg (m^3 m-3):} Soil moisture (10 - 40 cm underground).
\item \bold{SoilMoi100_200cm_tavg (m^3 m-3):} Soil moisture (100 - 200 cm underground).
\item \bold{SoilMoi40_100cm_tavg (m^3 m-3):} Soil moisture (40 - 100 cm underground).
\item \bold{SoilTemp00_10cm_tavg (K):} Soil temperature (0 - 10 cm underground).
\item \bold{SoilTemp10_40cm_tavg	(K):} Soil temperature (10 - 40 cm underground).
\item \bold{SoilTemp100_200cm_tavg (K):} Soil temperature (100 - 200 cm underground).
\item \bold{SoilTemp40_100cm_tavg (K):} Soil temperature (40 - 100 cm underground).
\item \bold{SWdown_f_tavg (W m-2):} Surface downward shortwave radiation.
\item \bold{SWE_inst (kg m-2):} Snow water equivalent.
\item \bold{Swnet_tavg	(W m-2):} Net shortwave radiation flux.
\item \bold{Tair_f_tavg (K):} Near surface air temperature.
\item \bold{Wind_f_tavg (m s-1):} Near surface wind speed.



}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
# 2. Extracting climate information
data <- region_ee \%>\% get_fldas(
  from = "2001-02-01", to = "2002-12-31",
  by = "month", band = "Qair_f_tavg", fun = "mean")
}
}
