% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghm.R
\name{get_ghm}
\alias{get_ghm}
\title{Extract Global Human Modification data}
\usage{
get_ghm(region)
}
\arguments{
\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract Global Human Modification data of the year 2016
}
