% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghsl.R
\name{get_ghsl}
\alias{get_ghsl}
\title{Extract Global Human Settlement Layer data}
\usage{
get_ghsl(region, scale = 100)
}
\arguments{
\item{region}{region and object sf.}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variable in km².
}
\description{
A function that extract Global Human Settlement Layer data from \bold{1975} to \bold{2030} each 5 years.
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region, id = 'distr' ,simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\% get_ghsl()
}
}
