% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_index.R
\name{get_ndvi}
\alias{get_ndvi}
\title{Extract of vegetation index}
\usage{
get_ndvi(year, region)
}
\arguments{
\item{year}{is date format for extract the variable}

\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract the values of vegetation
}
