% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa.R
\name{get_noaa}
\alias{get_noaa}
\title{Extract Night time light data}
\usage{
get_noaa(from, to, region, fun = "mean", scale = 100)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{region}{it's a feature collection.}

\item{fun}{function for extract statistic zonal (\bold{count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance}).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble with the new variables.
}
\description{
A function that extract a time series of night time light by \bold{year} from \bold{1992-2014}.
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region, id = 'distr' , simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\% get_noaa(
  from = "2001-01-01", to = "2003-01-01",fun = "max",scale = 1000)
}
}
