% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_od_data.R
\name{get_od_data}
\alias{get_od_data}
\title{Get census-like data in a origin-destination format}
\usage{
get_od_data(x, matrix = FALSE)
}
\arguments{
\item{x}{dataframe resulting from reading raw census data.}

\item{matrix}{logical. Should the output be in matrix form? Defaults to 'FALSE'.}
}
\value{
dataframe (or matrix) with origin-destination format.
}
\description{
Get census-like data in a origin-destination format
}
\examples{
\dontrun{
library(readxl)
library(fs)
raw_path <- path("data", "test")
report_path <- path(raw_path, "reporte.xlsx")
report_raw <- read_excel(report_path, col_names = FALSE)

od <- get_od_data(report_raw)
head(od)

od_matrix <- get_od_data(report_path, matrix = TRUE)
dim(od_matrix)
}
}
