% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop.R
\name{get_pop}
\alias{get_pop}
\title{Extract population data of WorldPop}
\usage{
get_pop(to, from, band, region, fun = "count")
}
\arguments{
\item{to, from}{the starting and final range of date.}

\item{band}{name of band.}

\item{region}{is a feature or feature collection}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance, first).}
}
\value{
a sf object with the new variables.
}
\description{
A function that extract a time series of the number of population by \bold{year}.
}
\details{
Name of some bands.
\itemize{
\item \bold{pupulation:} Estimated number of people residing in each grid cell.
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region, id = 'distr' , simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\% get_pop(
  to = "2001-01-01", from = "2002-12-31",
  band = "population", fun = "max")
}
}
