% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_index.R
\name{get_savi}
\alias{get_savi}
\title{Extract of vegetation index}
\usage{
get_savi(year, region)
}
\arguments{
\item{year}{is date format for extract the variable}

\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract the values of vegetation
}
\examples{
\dontrun{

library(lis)
library(rgee)
library(sf)
ee_Initialize()
region <- import_db("Peru_shp")[1, 1]
data <- get_evi(year = 2009, region = region)

}
}
