% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_so2.R
\name{get_so2}
\alias{get_so2}
\title{Extract Ozone data of Sentinel5}
\usage{
get_so2(from, to, band, region, fun = "max", scale = 1000)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{band}{name of band.}

\item{region}{is a feature or feature collection.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance, first).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a sf object with the new variables.
}
\description{
A function that extract a time series of ozone.
}
\details{
Name of some bands.
\itemize{
\item \bold{SO2_column_number_density:} SO2 vertical column density at ground level, calculated using the DOAS technique.
\item \bold{SO2_column_number_density_amf:} Weighted mean of cloudy and clear air mass factor (amf) weighted by intensity-weighted cloud fraction.
\item \bold{SO2_slant_column_number_density:} SO2 ring corrected slant column number density.
\item \bold{cloud_fraction:} Effective cloud fraction. See the Sentinel 5P L2 Input/Output Data Definition Spec, p.220.
\item \bold{sensor_azimuth_angle	:} Azimuth angle of the satellite at the ground pixel location (WGS84); angle measured East-of-North.
\item \bold{sensor_zenith_angle:} Zenith angle of the satellite at the ground pixel location (WGS84); angle measured away from the vertical.
\item \bold{solar_azimuth_angle:} Azimuth angle of the Sun at the ground pixel location (WGS84); angle measured East-of-North.
\item \bold{solar_zenith_angle:} Zenith angle of the satellite at the ground pixel location (WGS84); angle measured away from the vertical.
\item \bold{SO2_column_number_density_15km:} SO2 vertical column density at 15km, calculated using the DOAS technique.
 }
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
# 2. Extracting climate information
data <- region_ee \%>\% get_so2(
  from = "2019-02-01", to = "2019-12-31",
  band = "SO2_column_number_density", fun = "max")
}
}
