% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vegetation.R
\name{get_vegetation}
\alias{get_vegetation}
\title{Extract vegetation index of MODIS}
\usage{
get_vegetation(from, to, band, region, fun = "count", scale = 1000)
}
\arguments{
\item{to, from}{the starting and final range of date.}

\item{band}{name of bands.}

\item{region}{is a feature or feature collection.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median , min, mode , percentile , std, sum, variance, first).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variables.
}
\description{
A function that extract the values of vegetation by \bold{month} (2000-02-18T00:00:00Z - 2022-04-07T00:00:00).
}
\details{
Name of some bands.
\itemize{
\item \bold{NDVI:} Normalized Difference Vegetation Index.
\item \bold{EVI:} Enhanced Vegetation Index.
\item \bold{SAVI:} Soil Adjusted Vegetation Index.
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region, id = "distr", simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\% get_vegetation(
  from = "2001-02-01", to = "2002-12-31", band = "NDVI", fun = "max"
)
}
}
