% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getETP.R
\name{getETP}
\alias{getETP}
\title{Extract Evapotranspiration data of MODIS}
\usage{
getETP(year, region)
}
\arguments{
\item{year}{is date format for extract the variable}

\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract a ETP time series of MODIS
}
\examples{
\dontrun{
library(lis)
library(rgee)
library(sf)
ee_Initialize()

region <- st_read('../district.gpkg')
data <- getETP(year =2009,region=region)
}
}
