% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGhm.R
\name{getGhm}
\alias{getGhm}
\title{Extract Global Human Modification data}
\usage{
getGhm(region)
}
\arguments{
\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract Global Human Modification data of the year 2016
}
\examples{
\dontrun{
library(lis)
library(rgee)
library(sf)
ee_Initialize()

region <- st_read('../district.gpkg')
data <- getGhm(region=region)
}
}
