% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrbanlc.R
\name{getUrbanlc}
\alias{getUrbanlc}
\title{Extract urban area data of MODIS Landcover}
\usage{
getUrbanlc(year, region)
}
\arguments{
\item{year}{is date format for extract the variable}

\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract a time series of the urban area of MODIS Landcover
}
\examples{
\dontrun{
library(lis)
library(rgee)
library(sf)
ee_Initialize()

region <- st_read('../district.gpkg')
data <- getUrbanlc(year =2009,region=region)
}
}
