% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_db.R
\name{import_db}
\alias{import_db}
\title{Loads district data for Perú}
\usage{
import_db(dataset)
}
\arguments{
\item{dataset}{is a character specifying the dataset to request.

Current datasets available are:

- \bold{Peru_names}: Returns the name of the 1874 districts, 196 provinces and 25 regions of Peru as they are reported by the INEI in the REDATAM platform for the 2017 CENSUS.

- \bold{Peru_shp}: Returns the shapefile for the 1874 districts, 196 provinces and 25 regions of Peru.}
}
\description{
Function 'import_db' the specified dataset.
}
\examples{
# Import shapefiles of Peru
peru_shp <- import_db('Peru_shp')
head(peru_shp)

# Import names of distritcts
peru_names <- import_db('Peru_names')
head(peru_names)

}
