% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_innova.R
\name{innova_pal}
\alias{innova_pal}
\title{Return function to interpolate a lis color palette}
\usage{
innova_pal(palette = "ccvi", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in lis_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a lis color palette
}
\examples{
\dontrun{
library(innovar)
library(scales)
pal <- innova_pal("ccvi")(9)
show_col(pal)
}

}
