% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lis_algorithms.R
\name{lis_algorithms}
\alias{lis_algorithms}
\title{list_algorithms}
\usage{
lis_algorithms(category = "all")
}
\arguments{
\item{category}{type of algorithm}
}
\value{
a string vector
}
\description{
A function that show all the available functions of lis package.
}
\details{
category.
\itemize{
\item \bold{climate}
\item \bold{environment}
\item \bold{human_intervention}
\item \bold{vulnerability_index}
\item \bold{utils}
\item \bold{graph}
\item \bold{vulnerability_index}
\item \bold{admin}
}
}
