% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_lis.R
\docType{data}
\name{lis_colors}
\alias{lis_colors}
\title{Vector of custom colors}
\format{
An object of class \code{character} of length 55.
}
\usage{
lis_colors
}
\description{
Vector of custom colors
}
\keyword{datasets}
