% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_lis.R
\name{lis_cols}
\alias{lis_cols}
\title{Function to extract lis colors as hex codes}
\usage{
lis_cols(...)
}
\arguments{
\item{...}{Character names of lis_colors}
}
\description{
Function to extract lis colors as hex codes
}
