% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_lis.R
\name{lis_pal}
\alias{lis_pal}
\title{Return function to interpolate a lis color palette}
\usage{
lis_pal(palette = "ccvi", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in lis_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a lis color palette
}
