% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_lis.R
\docType{data}
\name{lis_palettes}
\alias{lis_palettes}
\title{Vector of custom palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
lis_palettes
}
\description{
Vector of custom palettes
}
\keyword{datasets}
