% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migration17raw}
\alias{migration17raw}
\title{Raw migration data from the Peru census of 2017}
\format{
A untidy data frame which resulted from reading an Excel file without
formatting or setting column names, and containing the information about migration
described above.
}
\source{
\url{https://censos2017.inei.gob.pe/redatam/}
}
\usage{
data(migration17raw)
}
\description{
The Institute of Statistics and Informatics (INEI, by its acronym in Spanish)
of Peru carried out the last census on 2017. This census data can be
queried in the \href{https://censos2017.inei.gob.pe/redatam/}{2017 Census RADATAM platform}.
}
\details{
This data set is obtained by reading the downloaded report generated by querying
the question \emph{Distrito o país donde vivía hace 5 años} (District or country
where you used to live 5 years ago) at a district level without setting column names
\code{readxl::read_excel(<report-filepath>, col_names = FALSE)}. The downloaded report
is an Excel file containing metadata about the query and a list for every district
with all the districts or countries given as a answer for this question and
their frequencies (number of people living in district A that used to live in
district B 5 years ago). Thus, we have 5-year migration information for all
the 1874 districts in Peru, which has to be processed to be in a tidy
origin-destination format.

The function \code{\link{get_od_data}} in this package aids to the processing
of this kind of reports from the Peru census of 2017.
}
\seealso{
\code{\link{get_od_data}}
}
\keyword{datasets}
