% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_string.R
\name{mirror_string}
\alias{mirror_string}
\title{mirror_string}
\usage{
mirror_string(string)
}
\arguments{
\item{string}{a string of the form a_b which needs to be mirrored in the for b_a. both a_b and b_a form a complete mirroed vector in the end.}
}
\value{
a mirrored string
}
\description{
Makes a mirror of strings of the form a_b
}
\examples{
mirror_string(c("2_1","1_3","10_2","100_4"))

}
