% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{module_coords}
\alias{module_coords}
\title{module_coords}
\usage{
module_coords(
  n_var = ncol(gen_att),
  which_neg = c("1_2", "4_5", "2_3", "20_4", "30_3"),
  dis_neg_know_ass = known_neg_dis_gen,
  coo_known_neg_ass = c("4_5", "2_3", "20_4", "30_3"),
  dis_pos_know_ass = known_pos_dis_gen,
  coo_known_pos_ass = c("3_9", "4_5", "6_8", "7_8"),
  by = 1e-04
)
}
\arguments{
\item{n_var}{number of the variables in the module  to be simulated}

\item{which_neg}{string with coordinates of the variables which are negativelly correlated.}

\item{dis_neg_know_ass}{Distances describing negative assocaitions known a priori}

\item{coo_known_neg_ass}{Coordinates that describe the positive association between two variables ex 1_2 indicates association between first and second variable}

\item{dis_pos_know_ass}{Distances describing positive assocaitions known a priori}

\item{coo_known_pos_ass}{Coordinates that describe the positive association between two variables ex 1_2 indicates association between first and second variable}

\item{by}{parameter like in the sample function used to generate randomly unknown positive and negative associations}
}
\value{
a list with an coords a data set named coords that describes the association to be considered between variables of a module to be simulated, and details to map such information into a matrix.
}
\description{
creates a data set named coords that describes the association to be considered between variables of a module to be simulated, and details to map such information into a matrix.
}
\examples{
module_coords(n_var=9,which_neg=c("2_1","6_2","1_9","1_6","2_3"),dis_neg_know_ass=c(-0.9300003, -0.9310004, -0.8684074),coo_known_neg_ass=c("2_1","6_2","1_9"),dis_pos_know_ass=c(0.5235988, 0.6435011, 1.1197695),coo_known_pos_ass=c("3_9","4_5","6_8"),by=0.001)
}
