% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Peru}
\alias{Peru}
\title{Loads district data for Peru}
\format{
A dataset with 1874 rows and 9 variables:
\describe{
  \item{ubigeo}{code of six numbers for each district}
  \item{reg.code}{code to each region}
  \item{reg}{name of each region}
  \item{prov.code}{code to each province}
  \item{prov}{name of each province}
  \item{distr.code}{code to each district}
  \item{distr}{name of each district}
  \item{capital}{name of the capital main}
}
}
\source{
\url{https://www.inei.gob.pe/}
}
\description{
A dataset containing the names of the 1874 districts, 196 provinces and 25 regions of Peru according INEI - 2017, additionally also is the limits administrative level 3 in a spatial vector format.
}
\keyword{data}
