% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_ee.R
\name{pto_to_ee}
\alias{pto_to_ee}
\title{Point to earth engine}
\usage{
pto_to_ee(x, id)
}
\arguments{
\item{x}{point sf object.}

\item{id}{the main id name in sf object.}
}
\description{
A function that transform points to featurecollection.
}
