% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_batch.R
\name{read_batch}
\alias{read_batch}
\title{Multiple import}
\usage{
read_batch(data_dir, extension = "csv", fun = readr::read_csv, env = TRUE, ...)
}
\arguments{
\item{argument_2}{Import multiple files}
}
\description{
https://www.gerkelab.com/blog/2018/09/import-directory-csv-purrr-readr/#'
}
\examples{
data_dir <- "~/Dropbox/Work/COVID19/Proyectos/MINEDU/LIS-MINEDU_Shared/Indicadores/3.2. Reapertura 2021/Output data/"

read_batch(data_dir = data_dir)
read_batch(data_dir = data_dir, extension = "dta", fun = haven::read_dta)
read_batch(data_dir = data_dir, extension = "dta", fun = haven::read_dta, env = F)

}
