% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_innova.R
\name{scale_color_innova}
\alias{scale_color_innova}
\title{Color scale constructor for lis colors}
\usage{
scale_color_innova(palette = "ccvi", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in lis_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for lis colors
}
\examples{
\dontrun{
library(ggplot2)
library(innovar)
# Default discrete palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point(size = 4) +
scale_color_innova()

# Default continuous palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
geom_point(size = 4, alpha = .6) +
scale_color_innova(discrete = FALSE)
}

}
