% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_innova.R
\name{scale_fill_innova}
\alias{scale_fill_innova}
\title{Fill scale constructor for lis colors}
\usage{
scale_fill_innova(palette = "ccvi", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in lis_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for lis colors
}
\examples{
\dontrun{
library(ggplot2)
library(innovar)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
geom_bar() +
theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
scale_fill_innova()
}

}
