% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sim_copula.R
\name{set_sim_copula}
\alias{set_sim_copula}
\title{set_sim_copula}
\usage{
set_sim_copula(
  d = 9,
  lower_tri_corr = soil_corr[lower.tri(soil_corr, diag = FALSE)],
  n_cont_var = ncol(soil_num),
  cont_var_par = list(list(lambda = 40), list(df = 30, ncp = 3), list(shape = 40, scale =
    10), list(rate = 30)),
  n_unique = length(unique(na.omit(soil_att_ex$siteid))),
  mar_cont_dists = c("pois", "chisq", "gamma", "exp"),
  var_names = colnames(soil_att_ex)
)
}
\arguments{
\item{d}{dimension of copula to be set. The same as the number of variables inside a module}

\item{lower_tri_corr}{lower triangular part of a valida correlation matrix}

\item{n_cont_var}{number of continuous variables inside the module and which will be described by specific marginal distributions}

\item{cont_var_par}{a list of lists containing names and values of the parameters of the}

\item{n_unique}{number of sites for which the simulations should be made}

\item{mar_cont_dists}{a vector of strings containing strings with the names of the marginal distributions}

\item{var_names}{name of the variables inside the moduls}
}
\value{
a list containing the copula settled for the simulations its mvdc component and the simulated continuos values. some of them will be siliced to generate categorical components
}
\description{
set_sim_copula
}
