% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_att_ex.r
\docType{data}
\name{soil_att_ex}
\alias{soil_att_ex}
\title{soil_att_ex soil attributes considered in small example of the package}
\format{
\subsection{\code{soil_att_ex}}{

A data frame containing small number of soil attributes
\describe{
\item{siteid}{Site IDs}
\item{Other attributes}{Other soil attributes measured}

...
}
}
}
\source{
<the author>
}
\usage{
soil_att_ex
}
\description{
A dataset containing attributes used to simulate soil information
}
\keyword{datasets}
