% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_att_innovar.r
\docType{data}
\name{soil_att_innovar}
\alias{soil_att_innovar}
\title{soil_att_ex soil attributes considered in small example of the package}
\format{
\subsection{\code{soil_att_innovar}}{

A data frame containing large number of soil attributes
\describe{
\item{siteid}{Site IDs}
\item{Other attributes}{Other soil attributes measured}

...
}
}
}
\source{
<the author>
}
\usage{
soil_att_innovar
}
\description{
A dataset containing attributes used to simulate soil information
}
\keyword{datasets}
