% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_root_matrix.R
\name{square_root_matrix}
\alias{square_root_matrix}
\title{square_root_matrix}
\usage{
square_root_matrix(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
x_sqrt the square root of a matrix.
}
\description{
Calculates the squared root of a matrix
}
\details{
the eigen vectors need of x need to  be positive.
}
\references{
Hands-On Matrix Algebra Using R chapter 11 describes root matrix
}
