% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_mtx.R
\name{sym_mtx}
\alias{sym_mtx}
\title{sym_mtx}
\usage{
sym_mtx(vector = 1:25, n_var = 5, var_names = c("v1", "v2", "v3", "v4", "v5"))
}
\arguments{
\item{vector}{a vector containing values to be used in the matrix creation}

\item{n_var}{number of variables}

\item{var_names}{name of the variables in the symmetric matrix}
}
\value{
a symmetric matrix
}
\description{
create a symetric matrix with n_var rows and n_var columns
}
\examples{
sym_mtx(vector=c(1:25), n_var=5,var_names=c("v1","v2","v3","v4","v5"))
}
