% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{toy_data a toy data set containing genomic and soil information to run models}
\format{
\subsection{\code{toy_data}}{

A data frame containing large number of soil attributes
\describe{
\item{trait_i}{genomic expression of trait i}
\item{soil attributes}{ soil attributes measured}

...
}
}
}
\source{
<the author>
}
\usage{
toy_data
}
\description{
A dataset containing genomic and soil information to run models
}
\keyword{datasets}
