% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_templates.R
\name{use_rmd_flatly}
\alias{use_rmd_flatly}
\title{Use Innovar Rmarkdown Flatly}
\usage{
use_rmd_flatly(file)
}
\arguments{
\item{file}{Name of the `.Rmd` file to create.}
}
\description{
This function populates the working directory with
files corresponding to the Innovar Rmarkdown Flatly.
}
\examples{
\dontrun{
library(innovar)
use_rmd_flatly("report_01.Rmd")
}
}
