% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_templates.R
\name{use_xaringan}
\alias{use_xaringan}
\title{Use Innovar Xaringan Template}
\usage{
use_xaringan(file)
}
\arguments{
\item{file}{Name of the `.Rmd` file to create.}
}
\description{
This function populates the working directory with
files corresponding to the xaringan template of the
package.
}
\examples{
\dontrun{
library(innovar)
use_xaringan("class_01.Rmd")
}
}
