% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability_index.R
\name{vulnerability_index}
\alias{vulnerability_index}
\title{Vulnerability Index}
\usage{
vulnerability_index(
  data,
  direct = NULL,
  inverse = NULL,
  table = NULL,
  weighted = TRUE,
  domains = "default",
  level = NULL,
  ordered = TRUE,
  method = "percent_rank",
  na.rm = FALSE
)
}
\arguments{
\item{na.rm}{}
}
\value{

}
\description{
It allows the observations to be ordered by different methods (default
  percentile rank), and in consideration of domains grouping indicators.
}
