% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllpaUtils_names.R
\name{AllpaUtils_names}
\alias{AllpaUtils_names}
\title{TIYLE}
\usage{
AllpaUtils_names(x)
}
\arguments{
\item{x}{vector of characters}
}
\description{
Function 'AllpaUtils_names' revisar: https://cran.r-project.org/web/packages/MazamaSpatialUtils/vignettes/MazamaSpatialUtils.html
}
\examples{
df

}
