% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpautils_names.R
\name{allpautils_names}
\alias{allpautils_names}
\title{Remove accents and hyphens in character vector}
\usage{
allpautils_names(string)
}
\arguments{
\item{string}{Characters vector}
}
\description{
This function will allow you to remove accents, hyphens and
  any strange characters that come from the UTF-8 encoding, with the intention
  of working with standardized names or texts.Review:
  https://cran.r-project.org/web/packages/MazamaSpatialUtils/vignettes/MazamaSpatialUtils.html
}
\examples{
string <- c("Perú", "Estados Unidos")
allpautils_names(string)

}
