% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpautils_names.R
\name{allpautils_names}
\alias{allpautils_names}
\title{TIYLE}
\usage{
allpautils_names(x)
}
\arguments{
\item{x}{vector of characters}
}
\description{
Function 'allpautils_names' revisar: https://cran.r-project.org/web/packages/MazamaSpatialUtils/vignettes/MazamaSpatialUtils.html
}
\examples{
df

}
