% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ee_median}
\alias{ee_median}
\title{Set of function to zonal statistic}
\usage{
ee_median(x, y, by = 1000)
}
\arguments{
\item{x}{image of type Image o Image Collection}

\item{y}{region of type Feacture o FeatureCollection}

\item{by}{a limit of pass

median}
}
\description{
Set of function to zonal statistic
}
