% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate.R
\name{get_climate}
\alias{get_climate}
\title{Extract climate data of TerraClimate}
\usage{
get_climate(from, to, by, band, region, fun = "count", scale = 1000)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{by}{two types of increment of the sequence by \bold{month} and \bold{year}.}

\item{band}{name of band.}

\item{region}{is a feature or feature collection.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance, first).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variables.
}
\description{
A function that extract a time series of climate variables (1958-01-01T00:00:00Z - 2021-12-01T00:00:00).
}
\details{
Name of some bands.
\itemize{
\item \bold{aet (mm):} Actual evapotranspiration, derived using a one-dimensional soil water balance model.
\item \bold{def (mm):} Climate water deficit, derived using a one-dimensional soil water balance model.
\item \bold{pdsi :} Palmer Drought Severity Index.
\item \bold{pet(mm):} Reference evapotranspiration (ASCE Penman-Montieth).
\item \bold{pr (mm):} Precipitation accumulation.
\item \bold{ro (mm):} Runoff, derived using a one-dimensional soil water balance model.
\item \bold{soil (mm):} Soil moisture, derived using a one-dimensional soil water balance model.
\item \bold{srad (W/m²):} Downward surface shortwave radiation.
\item \bold{swe (mm):} Snow water equivalent, derived using a one-dimensional soil water balance model.
\item \bold{tmmn (°C):} Minimum temperature.
\item \bold{tmmx (°C):} Maximum temperature.
\item \bold{vap (kPa):} Vapor pressure
\item \bold{vpd (kPa):} Vapor pressure deficit.
\item \bold{vs (m/s):} Wind-speed at 10m.
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
# 2. Extracting climate information
data <- region_ee \%>\% get_climate(
  from = "2001-02-01", to = "2002-12-31",
  by = "month", band = "tmmx", fun = "max")
}
}
