% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate.R
\name{get_climate}
\alias{get_climate}
\title{Extract climate data of TerraClimate}
\usage{
get_climate(to, from, by, band, region, fun = "count")
}
\arguments{
\item{to, from}{the starting and final range of date.}

\item{by}{two types of increment of the sequence by \bold{month} and \bold{year}.}

\item{band}{name of band.}

\item{region}{region and object sf.}

\item{fun}{function for extract statistic zonal ('count','kurtosis','max','mean','median','min','mode','percentile','std','sum','variance','first').}
}
\value{
a sf object with the new variables.
}
\description{
A function that extract a time series of climate variables.
}
\details{
Name of some bands.
\itemize{
\item \bold{aet:} Actual evapotranspiration, derived using a one-dimensional soil water balance model.
\item \bold{def:} Climate water deficit, derived using a one-dimensional soil water balance model.
\item \bold{pdsi:} Palmer Drought Severity Index.
\item \bold{pet:} Reference evapotranspiration (ASCE Penman-Montieth).
\item \bold{pr:} Precipitation accumulation.
\item \bold{ro:} Runoff, derived using a one-dimensional soil water balance model.
\item \bold{soil:} Soil moisture, derived using a one-dimensional soil water balance model.
\item \bold{srad:} Downward surface shortwave radiation.
\item \bold{swe:} Snow water equivalent, derived using a one-dimensional soil water balance model.
\item \bold{tmmn:} Minimum temperature.
\item \bold{tmmx:} Maximum temperature.
\item \bold{vap:} Vapor pressure
\item \bold{vpd:} Vapor pressure deficit.
\item \bold{vs:} Wind-speed at 10m.
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region , simplify = 1000)
# 2. Extracting climate information
data <- region_ee \%>\% get_climate(
  to = "2001-02-01", from = "2002-12-31",
  by = "month", band = "tmmx", fun = "max")
}
}
