% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_etp.R
\name{get_etp}
\alias{get_etp}
\title{Extract evapotranspiration data of Modis}
\usage{
get_etp(to, from, by, band, region, fun = "count")
}
\arguments{
\item{to, from}{is a string object,starting and final date.}

\item{by}{two types of increment of the sequence by \bold{month} and \bold{year}.}

\item{band}{name of band.}

\item{region}{region and object sf.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max,mean, median , min , mode , percentile, std, sum , variance, first).}
}
\value{
a sf object with the new variables.
}
\description{
A function that extract a ETP time series of MODIS
}
\details{
Name of some bands.
\itemize{
\item \bold{ET:} Total evapotranspiration.
\item \bold{LE:}  Average latent heat flux.
\item \bold{PET:} Total potential evapotranspiration.
\item \bold{PLE:} Average potential latent heat flux.
\item \bold{ET_QC:} Evapotranspiration quality control flags
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region,  id = 'distr' , simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\%
get_etp(to = "2001-02-01", from = "2003-12-31",
  by = "year", band = "ET", fun = "max")
}
}
