% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_etp.R
\name{get_etp}
\alias{get_etp}
\title{Extract evapotranspiration data of Modis}
\usage{
get_etp(from, to, band, region, fun = "count", scale = 1000)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{band}{name of band.}

\item{region}{region and object sf.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max,mean, median , min , mode , percentile, std, sum , variance, first).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variables.
}
\description{
A function that extract a ETP time series of MODIS by \bold{month} (2001-01-01T00:00:00Z - 2022-04-23T00:00:00).
}
\details{
Name of some bands.
\itemize{
\item \bold{ET (kg/m²):} Total evapotranspiration.
\item \bold{LE (J/m²):}  Average latent heat flux.
\item \bold{PET (kg/m²):} Total potential evapotranspiration.
\item \bold{PLE (j/m²):} Average potential latent heat flux.
\item \bold{ET_QC:} Evapotranspiration quality control flags
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region, id = "distr", simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\%
  get_etp(from = "2001-02-01", to = "2003-12-31", band = "ET", fun = "max")
}
}
