% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghm.R
\name{get_gHM}
\alias{get_gHM}
\title{Extract Global Human Modification data}
\usage{
get_gHM(region, fun = "count")
}
\arguments{
\item{region}{region and object sf.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median, min, mod, percentile, std, sum, variance, first).}
}
\value{
a sf object with the new variables.
}
\description{
A function that extract Global Human Modification data of the year \bold{2016}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region, id = 'distr' ,simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\%
get_gHM(
 fun = "max"
)
}
}
