% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghm.R
\name{get_ghm}
\alias{get_ghm}
\title{Extract Global Human Modification data}
\usage{
get_ghm(to, from, band, region, fun = "count")
}
\arguments{
\item{to, from}{the starting and final range of date.}

\item{band}{name of band.}

\item{region}{region and object sf.}

\item{fun}{function for extract statistic zonal ('count','kurtosis','max','mean','median','min','mode','percentile','std','sum','variance','first').}
}
\value{
a sf object with the new variables.
}
\description{
A function that extract Global Human Modification data of the year \bold{2016}.
}
\details{
Name of some bands.
\itemize{
\item \bold{gHM:} global Human Modification.
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region, simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\% get_ghm(
  to = "2001-02-01", from = "2002-12-31",
  by = "month", band = "tmmx", fun = "max")
}
}
