% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_no2.R
\name{get_no2}
\alias{get_no2}
\title{Extract Nitrogen Dioxide data of Sentinel5}
\usage{
get_no2(from, to, band, region, fun = "max", scale = 1000)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{band}{name of band.}

\item{region}{is a feature or feature collection.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance, first).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variables.
}
\description{
A function that extract a time series of nitrogen dioxide (2018-07-10T10:05:44Z - 2022-05-15T00:00:00).
}
\details{
Name of some bands.
\itemize{
\item \bold{NO2_column_number_density (mol/m²):} Total vertical column of NO2 (ratio of the slant column density of NO2 and the total air mass factor).
\item \bold{tropospheric_NO2_column_number_density (mol/m²):} tropospheric vertical column of NO2.
\item \bold{stratospheric_NO2_column_number_density (mol/m²):} stratospheric vertical column of NO2
\item \bold{NO2_slant_column_number_density (mol/m²):} NO2 slant column density
\item \bold{tropopause_pressure (Pa):} topopause pressure
\item \bold{absorbing_aerosol_index:} 	Aerosol index (at wavelengths 354/388, i.e. the OMI pair) from the AER_AI level 2 product. See Level 2 Algorithms - Aerosol Index.
\item \bold{cloud_fraction:} Effective cloud fraction. See the Sentinel 5P L2 Input/Output Data Definition Spec, p.220.
\item \bold{sensor_altitude (m):} Altitude of the satellite with respect to the geodetic sub-satellite point (WGS84).
\item \bold{sensor_azimuth_angle (degrees):} Azimuth angle of the satellite at the ground pixel location (WGS84); angle measured East-of-North.
\item \bold{sensor_zenith_angle (degrees):} Zenith angle of the satellite at the ground pixel location (WGS84); angle measured away from the vertical.
\item \bold{solar_azimuth_angle (degrees):} Azimuth angle of the Sun at the ground pixel location (WGS84); angle measured East-of-North.
\item \bold{solar_zenith_angle (degrees):} Zenith angle of the satellite at the ground pixel location (WGS84); angle measured away from the vertical.
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
# 2. Extracting climate information
data <- region_ee \%>\% get_no2(
  from = "2019-02-01", to = "2019-12-31",
  band = "NO2_column_number_density", fun = "max")
}
}
