% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop.R
\name{get_pop}
\alias{get_pop}
\title{Extract population data of WorldPop}
\usage{
get_pop(year, region)
}
\arguments{
\item{year}{is date format for extract the variable}

\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract a time series of the number of population
}
