% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop.R
\name{get_pop}
\alias{get_pop}
\title{Extract population data of WorldPop}
\usage{
get_pop(to, from, region, fun = "count")
}
\arguments{
\item{to, from}{are strings of starting and final range of date.}

\item{region}{is a feature collection.}

\item{fun}{function for extract statistic zonal (\bold{count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance}).}
}
\value{
a tibble with the new variables.
}
\description{
A function that extract a time series of the number of population by \bold{year}.
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region, id = 'distr' , simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\% get_pop(
  to = "2001", from = "2003",fun = "max")
}
}
