% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop.R
\name{get_pop}
\alias{get_pop}
\title{Extract population data of WorldPop}
\usage{
get_pop(from, to, region, fun = "count", scale = 100)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{region}{it's a feature collection.}

\item{fun}{function for extract statistic zonal (\bold{count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance}).}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble with the new variables.
}
\description{
A function that extract a time series of the number of population by \bold{year} (2000-01-01T00:00:00Z - 2021-01-01T00:00:00).
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
# 2. Extracting climate information
data <- region_ee \%>\% get_pop(
  from = "2001-01-01", to = "2003-01-01",fun = "max",scale = 100)
}
}
