% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_so2.R
\name{get_so2}
\alias{get_so2}
\title{Extract Ozone data of Sentinel5}
\usage{
get_so2(from, to, band, region, fun = "max", scale = 1000)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{band}{name of band.}

\item{region}{is a feature or feature collection.}

\item{fun}{function for extract statistic zonal (count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance, first).}

\item{scale}{A nominal scale in meters of the projection to work in.}

\item{by}{two types of increment of the sequence by \bold{month}.}
}
\value{
a sf object with the new variables.
}
\description{
A function that extract a time series of ozone.
}
\details{

}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
# 2. Extracting climate information
data <- region_ee \%>\% get_so2(
  from = "2019-02-01", to = "2019-12-31",
  band = "SO2_column_number_density", fun = "max")
}
}
