% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urban.R
\name{get_urban}
\alias{get_urban}
\title{Extract urban area data of MODIS Landcover}
\usage{
get_urban(to, from, region)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{region}{it's a feature or feature collection}
}
\value{
a tibble object with the new variables
}
\description{
A function that extract a time series of the urban area of MODIS Landcover
}
\examples{
\dontrun{

library(lis)
library(rgee)
library(sf)
ee_Initialize()
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region, id = 'distr' ,simplify = 1000)
data <- get_climate(to = '2008', from = '2010', region = region)

}
}
