% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urban.R
\name{get_urban}
\alias{get_urban}
\title{Extract urban area data of MODIS Landcover}
\usage{
get_urban(from, to, region, scale = 1000)
}
\arguments{
\item{to, from}{it's a string object,starting and final date.}

\item{region}{it's a feature or feature collection}

\item{scale}{A nominal scale in meters of the projection to work in.}
}
\value{
a tibble object with the new variable in km2
}
\description{
A function that extract a time series of the urban area of MODIS Landcover (2001-01-01T00:00:00Z - 2020-01-01T00:00:00)
}
\examples{
\dontrun{

library(innovar)
library(rgee)
library(sf)
ee_Initialize()
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region, id = 'distr' ,simplify = 1000)
data <- get_urban(from = '2008-01-01', to = '2010-01-01', region = region)

}
}
