% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urban.R
\name{get_urban}
\alias{get_urban}
\title{Extract urban area data of MODIS Landcover}
\usage{
get_urban(year, region)
}
\arguments{
\item{year}{is date format for extract the variable}

\item{region}{is a sf object}
}
\value{
a sf object with the new variables
}
\description{
A function that extract a time series of the urban area of MODIS Landcover
}
\examples{
\dontrun{

library(lis)
library(rgee)
library(sf)
ee_Initialize()
region <- import_db("Peru_shp")[1,1]
data <- get_climate(year = 2009, region = region)

}
}
