% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vegetation.R
\name{get_vegetation}
\alias{get_vegetation}
\title{Extract of vegetation index}
\usage{
get_vegetation(to, from, by, band, region, fun = "count")
}
\arguments{
\item{to, from}{the starting and final range of date.}

\item{by}{two types of increment of the sequence by \bold{month} and \bold{year}.}

\item{band}{name of band.}

\item{region}{region and object sf.}

\item{fun}{function for extract statistic zonal ('count','kurtosis','max','mean','median','min','mode','percentile','std','sum','variance','first').}
}
\value{
a sf object with the new variables.
}
\description{
A function that extract the values of vegetation.
}
\details{
Name of some bands.
\itemize{
\item \bold{NDVI:} Normalized Difference Vegetation Index.
\item \bold{EVI:} Enhanced Vegetation Index.
}
}
\examples{
\dontrun{

library(tidyverse)
library(rgee)
library(lis)
library(sf)
ee_Initialize()

# 1. Reading a sf object
region <- import_db("Peru_shp")
region_ee <- pol_as_ee(region, simplify = 1000)

# 2. Extracting climate information
data <- region_ee \%>\% get_vegetation(
  to = "2001-02-01", from = "2002-12-31",
  by = "month", band = "tmmx", fun = "max")
}
}
