% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_db.R
\name{import_db}
\alias{import_db}
\title{Loads district data for Perú}
\usage{
import_db(dataset)
}
\arguments{
\item{dataset}{is a character specifiying the dataset to request.

Current datasets available are:

- "Peru_names": Returns the name of the 1874 districts, 196 provinces and 25 regions of Perú as they are reported by the INEI in the REDATAM platform for the 2017 CENSUS.

- "Peru_shp": Returns the shapefile for thethe 1874 districts, 196 provinces and 25 regions of Perú}
}
\description{
Function 'import_db' the specified dataset.
}
\examples{
df <- data.frame(
  reg = c(
    "LIMA", "CALLAO", "CAJAMARCA", "AMAZONAS", "SAN MARTIN", "HUANUCO",
    "PASCO", "JUNIN", "CUSCO", "PUNO", "APURIMAC", "AYACUCHO",
    "HUANCAVELICA", "TUMBES", "PIURA", "LAMBAYEQUE", "LA LIBERTAD",
    "ANCASH", "ICA", "AREQUIPA", "TACNA", "MOQUEGUA", "LORETO", "UCAYALI",
    "MADRE DE DIOS"
  ),
  stringsAsFactors = FALSE
)
}
