% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Peru}
\alias{Peru}
\title{Peru names and spatial}
\format{
A dataset with 1874 rows and 9 variables:

\describe{
  \item{\bold{ubigeo}}{Code of six numbers for each district}
  \item{\bold{reg.code}}{Code to each region}
  \item{\bold{reg}}{Name of each region}
  \item{\bold{prov.code}}{Code to each province}
  \item{\bold{prov}}{Name of each province}
  \item{\bold{distr.code}}{Code to each district}
  \item{\bold{distr}}{Name of each district}
  \item{\bold{capital}}{Name of the capital main}
}
}
\source{
\url{https://www.inei.gob.pe/}
}
\usage{
data(Peru)
}
\description{
A dataset containing the names of the 1874 districts, 196 provinces and 25 regions of Peru according INEI - 2017, additionally also is the limits administrative level 3 in a spatial vector format.
}
\keyword{datasets}
