% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_ee.R
\name{pol_as_ee}
\alias{pol_as_ee}
\title{Polygon to earth engine}
\usage{
pol_as_ee(x, id, simplify = 500)
}
\arguments{
\item{x}{polygon sf object.}

\item{id}{the main id name in sf object.}

\item{simplify}{number to simplify the sf polygon  keeping its topology}
}
\value{
feature collection
}
\description{
A function that transform polygon to featurecollection.
}
\examples{
\dontrun{
library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()
# 1. Reading a sf object
data("Peru")
region <- Peru \%>\% filter(reg == 'LIMA')
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
class(region_ee)
}
}
