% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_innova.R
\name{show_pal}
\alias{show_pal}
\title{Show a set of palette color Lab}
\usage{
show_pal(name = "all", n = 5, rev = TRUE, ...)
}
\arguments{
\item{name}{Character name of palette in lis_palettes}

\item{n}{Integer, number of colors}

\item{rev}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to seecol()}
}
\description{
Show a set of palette color Lab
}
\details{
Name of color palette available
\itemize{
\item \bold{ccvi}
\item \bold{npr}
\item \bold{blmbrg}
\item \bold{ctp}
\item \bold{jama}
\item \bold{mlobo}
\item \bold{btran}
\item \bold{nasa}
\item \bold{politico}
\item \bold{mortality}
}
}
\examples{
\dontrun{
library(innovar)
# show_pal(name = "nasa",rev = TRUE, n = 5)
show_pal()
}
}
