% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability_index.R
\name{vulnerability_index}
\alias{vulnerability_index}
\title{Vulnerability Index}
\usage{
vulnerability_index(
  data,
  direct = NULL,
  inverse = NULL,
  table = NULL,
  weighted = TRUE,
  domains = "default",
  level = NULL,
  ordered = TRUE,
  method = "percent_rank",
  na.rm = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the variables that will be analyzed.}

\item{direct}{A vector indicating the name of the variables that contribute
directly to the vulnerability index.}

\item{inverse}{A vector indicating the name of the variables that contribute
inversely to the vulnerability index.}

\item{table}{A dataframe indicating the variables "codename", "Direccion",
"Dominio" and "Escala_Geografica" as a minimum. If this argument is specified,
it is no longer necessary to specify the direct and inverse arguments.}

\item{weighted}{A logical value indicating whether the "Dominio" information is
used for the index calculation.}

\item{domains}{A vector indicating the names of the domains grouping the variables
of interest. If the argument "table" containing the variable "Dominio" is used,
only "default" is indicated in this argument.}

\item{level}{A vector indicating the level of analysis. It currently supports 3
values:"departamental", "provincial", and "distrital".}

\item{ordered}{A logical value indicating whether the cases analyzed will be
ordered from highest to lowest.}

\item{method}{A vector indicating the analysis method used for the index.
Currently only "percent_rank" is supported.}

\item{na.rm}{A logical value indicating whether cases where missing values are
eliminated.}
}
\value{
A [tibble::tibble()] containing the variables analyzed by the function.
}
\description{
It allows the observations to be ordered by different methods (default
  percentile rank), and in consideration of domains grouping indicators.
}
\seealso{
[dplyr::percent_rank()].
}
